requirements.txt
streamlit
torch
diffusers
transformers
accelerate
safetensors
Pillow
numpy
